+++ modules/template/elements/HtmlElement.php
@@ -65,6 +65,14 @@ class HtmlElement extends BaseElementContent implements \Stringable
     /**
      * @inheritdoc
      */
+    /**
+     * @inheritdoc
+     */
+    public static function isInlineEditingEnabledByDefault(): bool
+    {
+        return true;
+    }
+
     public function __toString(): string
     {
         return (string) $this->purify($this->content);