+++ docs/CHANGELOG.md
@@ -9,6 +9,8 @@ Changelog
 - Fix: Edit buttons of template elements were hidden behind the structure view when it overlapped them
 - Fix: Edit button of a single container item was covered by the add button of its container or disappeared when moving the mouse over it
 - Enh: Manage the items of container elements (edit, sort, add, delete) in the "Edit elements" dialog of a template instance
+- Enh: Inline editing of template elements directly on the page, can be enabled per element in the template administration
+- Enh: Clearer option "Allow inline HTML" (formerly "Is inline text") for text elements with a description of the encoding
 
 1.12.21 (September 8, 2026)
 ---------------------------